Changes to searchbox color and treeitems
Added back color for searchbox when it's not empty.
Added ability to clone with the same name if you're cloning to a
different mod.
Show at least mod name treeitem to allow adding aliases from scratch
when the mod is empty, and to confirm to the user that a new mod has
been created.